pyproject.toml
[tool.poetry]
name = "thick-denim"
version = "0.1.0"
description = "Thick-Denim is a python tool that allows querying and cross-referencing GitHub and Jira and processing data into metrics. Provides DSLs to describe queries, references and data processing."
authors = ["Gabriel Falcão <gabriel@nacaolivre.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
jira = "^2.0"
pendulum = "^2.0"
requests = "^2.22"
humanfriendly = "^4.18"
click = "^7.0"
"ruamel.yaml" = "^0.16.5"
pync = "^2.0"

[tool.poetry.dev-dependencies]
sure = "^1.4"
coverage = "^4.5"
doc8 = "^0.8.0"
flake8 = "^3.7"
freezegun = "^0.3.12"
httpretty = "^0.9.6"
ipdb = "^0.12.2"
mccabe = "^0.6.1"
mock = "^3.0"
nose = "^1.3"
nose-timer = "^0.7.5"
nose-watch = "^0.9.2"
parameterized = "^0.7.0"
rednose = "^1.3"
sphinx-click = "^2.3"
sphinx-rtd-theme = "^0.4.3"
vcrpy = "^2.1"
black = {version = "^18.3-alpha.0",allows-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
